refactor: creation of the query.js file and removing the query logic from the script.js file
